fix: merge both envs into one file
