+++ src/chemenv/modal_app.py
@@ -1,5 +1,13 @@
 import os
-from modal import App
+from modal import App, Volume
+from typing import Any
+from chemenv.tools.rxn_schema_processing import (
+    decimer_image,
+    _decimer_extractor,
+    rxnscribe_image,
+    _molscribe_extractor,
+    _rxnscribe_extractor,
+)
 from chemenv.tools.cheminformatics import (
     rdkit_image,
     mendeleev_image,
@@ -56,13 +64,81 @@ from chemenv.tools.rxn_utils import (
     _canonicalize_rxn_smiles,
 )
 
+MINUTES = 60  # 60 seconds
+
 chemenv_name = os.getenv("CHEMENV_NAME", "")
 if chemenv_name and not chemenv_name.startswith("-"):
     chemenv_name = f"-{chemenv_name}"
 
+app_name = os.getenv("RXNENV_NAME", "")
+if app_name and not app_name.startswith("-"):
+    app_name = f"-{app_name}"
+
+# Define the volume to safe checkpoints
+hf_cache_vol = Volume.from_name("huggingface-cache", create_if_missing=True)
+
+# Create the app
+rxn_app = App(f"rxnenv{app_name}")
 
 # Create the app
 app = App(f"chemenv{chemenv_name}")
+
+
+@rxn_app.function(image=decimer_image)
+def molecule_image_extraction_decimer(*args, **kwargs) -> str:
+    """
+    Predicts the SMILES from the image using DECIMER models.
+    https://github.com/Kohulan/DECIMER-Image_Transformer
+
+    Args:
+        image: str: Path to the image file.
+
+    Returns:
+        str: SMILES predicted.
+    """
+    return _decimer_extractor(*args, **kwargs)
+
+
+@rxn_app.function(
+    image=rxnscribe_image,
+    gpu="A10G",
+    volumes={
+        "/root/.cache/huggingface": hf_cache_vol,
+    },
+)
+def molecule_image_extraction_molscribe(*args, **kwargs) -> dict[str, Any]:
+    """
+    Extracts the reaction scheme from the image using MolScribe model.
+    https://github.com/thomas0809/MolScribe
+
+    Args:
+        image: bytes: Image in bytes.
+
+    Returns:
+        dict[str, Any]: Dictionary containing the reaction schema.
+    """
+    return _molscribe_extractor(*args, **kwargs)
+
+
+@rxn_app.function(
+    image=rxnscribe_image,
+    gpu="A10G",
+    volumes={
+        "/root/.cache/huggingface": hf_cache_vol,
+    },
+)
+def rxn_schema_extraction(*args, **kwargs) -> list[dict]:
+    """
+    Extracts the reaction scheme from the image using RxnScribe model.
+    https://github.com/thomas0809/RxnScribe
+
+    Args:
+        image: bytes: Image in bytes.
+
+    Returns:
+        list[dict]: List of dictionaries containing the reaction schema.
+    """
+    return _rxnscribe_extractor(*args, **kwargs)
 
 
 @app.function(image=rdkit_image)
